BCB Spring 2014

+++ bcb_app/src/com/bangalore/barcamp/widgets/CircularImageView.java
@@ -287,7 +287,7 @@ public class CircularImageView extends ImageView {
 				}
 				synchronized (fileWriterMonitor) {
 
-					if (!file.exists()) {
+					if (!file.exists() && bitmap != null) {
 						FileOutputStream fos;
 						try {
 							fos = new FileOutputStream(file);
